requirements.txt
discord.py
tqdm
python-dotenv
openai
deep_translator
langdetect
fastapi[all]